+++ yash-env/src/system.rs
@@ -142,6 +142,7 @@ use self::select::SignalStatus;
 pub use self::shared::SharedSystem;
 pub use self::signal::{
     CaughtSignals, Disposition, GetSigaction, SendSignal, Sigaction, Sigmask, SigmaskOp, Signals,
+    Sigset,
 };
 pub use self::sysconf::{ShellPath, Sysconf};
 pub use self::terminal::{Isatty, TcGetPgrp, TcSetPgrp};